fix(agent): 修复线程池复用及链路追踪功能
- P0-4: 改用共享缓存线程池替代每次新建单线程池，防止高并发下线程泄漏
- 超时后显式取消 Future 任务，避免子线程继续运行占用资源
- 添加 @PreDestroy 注解优雅关闭线程池，保证生命周期管理
- P2-1: 完善 MDC 上下文保存与恢复，支持嵌套调用时多 key 透传与恢复
- P2-8: 引入带上限的指数退避与随机抖动，降低重试冲击与等待爆炸
- AgentServiceImpl 引入 AgentTracer 进行全链路追踪，支持异常记录与链路结束
- 异步执行改用显式捕获并恢复 TenantContext，防止异步丢失租户上下文
- 流式执行中使用复合监听器和信号守护，避免重复触发 onError/onComplete 信号
- 任务编码改用雪花算法生成，避免并发重复导致唯一索引冲突
- AlertProperties 新增飞书和短信通知通道配置，支持多通知方式扩展

+++ ydsz-pmis-backend/ydsz-pmis-message/src/main/java/com/njydsz/pmis/message/controller/MessageController.java
@@ -134,8 +134,8 @@ public class MessageController {
     @PrePermission(PermissionCodes.MESSAGE_LOG_VIEW)
     @GetMapping("/batch/{batchId}/progress")
     public Result<Page<MsgLogDO>> batchProgress(@PathVariable String batchId,
-                                                @RequestParam(defaultValue = "1") int page,
-                                                @RequestParam(defaultValue = "20") int size) {
+                                                @RequestParam(defaultValue = "1") long page,
+                                                @RequestParam(defaultValue = "20") long size) {
         MessageLogQueryDTO query = new MessageLogQueryDTO();
         query.setBizId(batchId);
         query.setPage(page);
